refactor(controller): move fn GetFCertificateFromKongCert PluginString
